Day100/requirements.txt
Flask>=3.0
python-dotenv
openai>=1.0.0